Move ChangesetLocalId and ChangeAtomId
