+++ experimental/dds/tree2/src/core/rebase/index.ts
@@ -11,6 +11,8 @@ export {
 	GraphCommit,
 	RevisionTag,
 	RevisionTagSchema,
+	ChangesetLocalId,
+	ChangeAtomId,
 	SessionId,
 	SessionIdSchema,
 } from "./types";
