+++ experimental/dds/tree2/src/core/rebase/types.ts
@@ -6,7 +6,7 @@
 import { assert } from "@fluidframework/common-utils";
 import { isStableId } from "@fluidframework/container-runtime";
 import { StableId } from "@fluidframework/runtime-definitions";
-import { brandedStringType, generateStableId } from "../../util";
+import { Brand, brandedStringType, generateStableId } from "../../util";
 import { ReadonlyRepairDataStore } from "../repair";
 
 /**
@@ -23,6 +23,33 @@ export const SessionIdSchema = brandedStringType<SessionId>();
 // TODO: These can be compressed by an `IdCompressor` in the future
 export type RevisionTag = StableId;
 export const RevisionTagSchema = brandedStringType<StableId>();
+
+/**
+ * An ID which is unique within a revision of a `ModularChangeset`.
+ * A `ModularChangeset` which is a composition of multiple revisions may contain duplicate `ChangesetLocalId`s,
+ * but they are unique when qualified by the revision of the change they are used in.
+ * @alpha
+ */
+export type ChangesetLocalId = Brand<number, "ChangesetLocalId">;
+
+/**
+ * A globally unique ID for an atom of change, or a node associated with the atom of change.
+ * @alpha
+ *
+ * @privateRemarks
+ * TODO: Rename this to be more general.
+ */
+export interface ChangeAtomId {
+	/**
+	 * Uniquely identifies the changeset within which the change was made.
+	 * Only undefined when referring to an anonymous changesets.
+	 */
+	readonly revision?: RevisionTag;
+	/**
+	 * Uniquely identifies, in the scope of the changeset, the change made to the field.
+	 */
+	readonly localId: ChangesetLocalId;
+}
 
 /**
  * @returns a `RevisionTag` from the given string, or fails if the string is not a valid `RevisionTag`
