+++ experimental/dds/tree2/src/feature-libraries/index.ts
@@ -97,7 +97,6 @@ import * as SequenceField from "./sequence-field";
 export { SequenceField };
 
 export {
-	ChangesetLocalId,
 	idAllocatorFromMaxId,
 	isNeverField,
 	ModularEditBuilder,
@@ -137,7 +136,6 @@ export {
 	NodeExistsConstraint,
 	NodeExistenceState,
 	BrandedFieldKind,
-	ChangeAtomId,
 } from "./modular-schema";
 
 export {
