+++ experimental/dds/tree2/src/feature-libraries/modular-schema/index.ts
@@ -20,7 +20,6 @@ export {
 	idAllocatorFromMaxId,
 	setInCrossFieldMap,
 } from "./crossFieldQueries";
-export { ChangesetLocalId, ChangeAtomId } from "./modularChangeTypes";
 export { ChangesetLocalIdSchema, EncodedChangeAtomId } from "./modularChangeFormat";
 export {
 	FieldKind,
