+++ experimental/dds/tree2/src/feature-libraries/modular-schema/modularChangeTypes.ts
@@ -3,32 +3,8 @@
  * Licensed under the MIT License.
  */
 
-import { FieldKey, FieldKindIdentifier, RevisionTag } from "../../core";
+import { ChangesetLocalId, FieldKey, FieldKindIdentifier, RevisionTag } from "../../core";
 import { Brand } from "../../util";
-
-/**
- * An ID which is unique within a revision of a `ModularChangeset`.
- * A `ModularChangeset` which is a composition of multiple revisions may contain duplicate `ChangesetLocalId`s,
- * but they are unique when qualified by the revision of the change they are used in.
- * @alpha
- */
-export type ChangesetLocalId = Brand<number, "ChangesetLocalId">;
-
-/**
- * A globally unique ID for an atom of change, or a node associated with the atom of change.
- * @alpha
- */
-export interface ChangeAtomId {
-	/**
-	 * Uniquely identifies the changeset within which the change was made.
-	 * Only undefined when referring to an anonymous changesets.
-	 */
-	readonly revision?: RevisionTag;
-	/**
-	 * Uniquely identifies, in the scope of the changeset, the change made to the field.
-	 */
-	readonly localId: ChangesetLocalId;
-}
 
 /**
  * @alpha
