+++ experimental/dds/tree2/src/test/feature-libraries/sequence-field/sequenceFieldEditor.spec.ts
@@ -5,11 +5,8 @@
 
 import { strict as assert } from "assert";
 import { jsonString } from "../../../domains";
-import {
-	ChangesetLocalId,
-	SequenceField as SF,
-	singleTextCursor,
-} from "../../../feature-libraries";
+import { ChangesetLocalId } from "../../../core";
+import { SequenceField as SF, singleTextCursor } from "../../../feature-libraries";
 import { brand } from "../../../util";
 import { deepFreeze } from "../../utils";
 import { TestChange } from "../../testChange";
