+++ experimental/dds/tree2/src/test/feature-libraries/sequence-field/sequenceFieldToDelta.spec.ts
@@ -11,9 +11,9 @@ import {
 	ITreeCursorSynchronous,
 	TreeSchemaIdentifier,
 	mintRevisionTag,
+	ChangesetLocalId,
 } from "../../../core";
 import {
-	ChangesetLocalId,
 	FieldChange,
 	FieldKinds,
 	NodeChangeset,
